move ToggleColorMode to shared-theme
