+++ docs/data/material/getting-started/templates/shared-theme/AppTheme.tsx
@@ -10,29 +10,32 @@ import { dataDisplayCustomizations } from './customizations/dataDisplay';
 import { feedbackCustomizations } from './customizations/feedback';
 import { navigationCustomizations } from './customizations/navigation';
 import { surfacesCustomizations } from './customizations/surfaces';
-import {
-  colorSchemes,
-  cssVariables,
-  typography,
-  shadows,
-  shape,
-} from './themePrimitives';
+import { colorSchemes, typography, shadows, shape } from './themePrimitives';
+
+interface AppThemeProps {
+  children: React.ReactNode;
+  /**
+   * This is for the docs site. You can ignore it or remove it.
+   */
+  disableCustomTheme?: boolean;
+  themeComponents?: ThemeOptions['components'];
+}
 
 export default function AppTheme({
   children,
   disableCustomTheme,
   themeComponents,
-}: {
-  children: React.ReactNode;
-  disableCustomTheme?: boolean;
-  themeComponents?: ThemeOptions['components'];
-}) {
+}: AppThemeProps) {
   const theme = React.useMemo(() => {
     return disableCustomTheme
       ? {}
       : createTheme({
-          colorSchemes,
-          cssVariables,
+          // For more details about CSS variables configuration, see https://mui.com/material-ui/customization/css-theme-variables/configuration/
+          cssVariables: {
+            colorSchemeSelector: 'data-mui-color-scheme',
+            cssVarPrefix: 'template',
+          },
+          colorSchemes, // Recently added in v6 for building light & dark mode app, see https://mui.com/material-ui/customization/palette/#color-schemes
           typography,
           shadows,
           shape,
