+++ docs/data/material/getting-started/templates/shared-theme/ToggleColorMode.tsx
@@ -1,0 +1,23 @@
+import * as React from 'react';
+import { useColorScheme } from '@mui/material/styles';
+import MenuItem from '@mui/material/MenuItem';
+import Select from '@mui/material/Select';
+
+export default function ToggleColorMode() {
+  const { mode, setMode } = useColorScheme();
+  if (!mode) {
+    return null;
+  }
+  return (
+    <Select
+      data-template-mode-trigger=""
+      value={mode}
+      onChange={(e) => setMode(e.target.value as 'system' | 'light' | 'dark')}
+      sx={{ position: 'fixed', top: '1rem', right: '1rem' }}
+    >
+      <MenuItem value="system">System</MenuItem>
+      <MenuItem value="light">Light</MenuItem>
+      <MenuItem value="dark">Dark</MenuItem>
+    </Select>
+  );
+}
