+++ docs/data/material/getting-started/templates/sign-in/SignIn.tsx
@@ -16,7 +16,7 @@ import { styled } from '@mui/material/styles';
 import ForgotPassword from './ForgotPassword';
 import { GoogleIcon, FacebookIcon, SitemarkIcon } from './CustomIcons';
 import AppTheme from '../shared-theme/AppTheme';
-import ToggleColorMode from './ToggleColorMode';
+import ToggleColorMode from '../shared-theme/ToggleColorMode';
 
 const Card = styled(MuiCard)(({ theme }) => ({
   display: 'flex',
